+++ include/GeneralClasses/PhysicalConstants.h
@@ -9,23 +9,23 @@ class PhysicalConstants {
 		/**
 		 * @brief c is the light's velocity.
 		 */
-        constexpr static long double c = 299792458;
+        constexpr static double c = 299792458;
 		/**
 		 * @brief h is the Planck's constant.
 		 */
-        constexpr static long double h = 6.62606957E-34;
+        constexpr static double h = 6.62606957E-34;
 		/**
 		 * @brief lambda is the light's wavelength.
 		 */
-		constexpr static long double lambda = 1550E-9;
+		constexpr static double lambda = 1550E-9;
 		/**
 		 * @brief freq is the light's frequency;
 		 */
-        constexpr static long double freq = 193.4E12;
+        constexpr static double freq = 193.4E12;
         /**
          * @brief BRef is the reference bandwidth.
          */
-        static constexpr long double BRef = 12.5E9;
+        static constexpr double BRef = 12.5E9;
         /**
          * @brief numPolarizations is used to choose whether
          * one or two polarizations are used to transmit the signal.
