+++ include/RWA/RoutingAlgorithms/LengthOccupationRoutingAvailability.h
@@ -7,7 +7,7 @@ class LengthOccupationRoutingAvailability : public DijkstraRoutingAlgorithm {
   public:
     LengthOccupationRoutingAvailability(std::shared_ptr<Topology> T);
 
-    long double get_Cost(std::weak_ptr<Link> link, std::shared_ptr<Call> C);
+    double get_Cost(std::weak_ptr<Link> link, std::shared_ptr<Call> C);
     void load() {}
 };
