+++ include/SimulationTypes/Simulation_NetworkLoad.h
@@ -23,8 +23,8 @@ class Simulation_NetworkLoad : public SimulationType {
     bool hasSimulated;
     bool hasLoaded;
 
-    long double NumCalls;
-    long double NetworkLoadMin, NetworkLoadMax, NetworkLoadStep;
+    double NumCalls;
+    double NetworkLoadMin, NetworkLoadMax, NetworkLoadStep;
 
     RoutingAlgorithm::RoutingAlgorithms
     Routing_Algorithm;
