+++ include/Structure/Topology.h
@@ -56,15 +56,15 @@ class Topology {
                                  Node::NodeType = Node::TransparentNode,
                                  Node::NodeArchitecture = Node::SwitchingSelect, int NumReg = 0);
     std::weak_ptr<Link> add_Link(std::weak_ptr<Node> Origin,
-                                 std::weak_ptr<Node> Destination, long double Length);
+                                 std::weak_ptr<Node> Destination, double Length);
 
     static std::shared_ptr<Topology> create_DefaultTopology(DefaultTopologies);
 
     void save(std::string TopologyFileName);
 
-    long double get_LengthLongestLink();
+    double get_LengthLongestLink();
   private:
-    long double LongestLink;
+    double LongestLink;
 };
 
 #endif // TOPOLOGY_H
