+++ src/RWA/RoutingAlgorithms/LengthOccupationRoutingAvailability.cpp
@@ -6,9 +6,9 @@ LengthOccupationRoutingAvailability::LengthOccupationRoutingAvailability(
 
 }
 
-long double LengthOccupationRoutingAvailability::get_Cost(
+double LengthOccupationRoutingAvailability::get_Cost(
     std::weak_ptr<Link> link, std::shared_ptr<Call>) {
     return 1 +
            (link.lock()->Length / T->get_LengthLongestLink()) +
-           (link.lock()->get_Availability() / (long double) Link::NumSlots);
+           (link.lock()->get_Availability() / (double) Link::NumSlots);
 }
